refactor: AccessGroupsDropdown typescript rewrite
